Use pluginv2 in authz plugins.
pluginstore provides the abstraction for legacy (pluginv1) and new
plugin model (pluginv2). In pluginv2, plugins are installed from the
registry and loaded right away (no lazy loading). Make authz capable of
using pluginv2.
